+++ core/src/main/java/com/graphhopper/routing/util/CarFlagEncoder.java
@@ -93,7 +93,7 @@ public class CarFlagEncoder extends AbstractFlagEncoder
     }
 
     @Override
-    public int isAllowed( OSMWay way )
+    public long isAllowed( OSMWay way )
     {
         String highwayValue = way.getTag("highway");
         if (highwayValue == null)
@@ -131,12 +131,12 @@ public class CarFlagEncoder extends AbstractFlagEncoder
     }
 
     @Override
-    public int handleWayTags( int allowed, OSMWay way )
+    public long handleWayTags( long allowed, OSMWay way )
     {
         if ((allowed & acceptBit) == 0)
             return 0;
 
-        int encoded;
+        long encoded;
         if ((allowed & ferryBit) == 0)
         {
             String highwayValue = way.getTag("highway");
@@ -178,7 +178,7 @@ public class CarFlagEncoder extends AbstractFlagEncoder
     }
 
     @Override
-    public int analyzeNodeTags( OSMNode node )
+    public long analyzeNodeTags( OSMNode node )
     {
 
         // absolute barriers always block
